requirements.txt
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.9
python-telegram-bot==13.5
requests==2.25.1
youtube-dl==2021.1.24.1
